Add files via upload
The most common issue that arises during the implementation of YOLOv8 on a custom dataset is the file paths. You must provide the correct file paths, otherwise YOLOv8 will throw an error.

To avoid this issue, you can follow the steps given on the Roboflow website to annotate your dataset. The website provides detailed instructions on how to create a text file that lists the relative paths to all of your images and labels. This text file can then be referenced in the YOLOv8 configuration file.